source/src/batch-evaluation/molecular-unfolding/image/lambda-image/device-available-check/requirements.txt
amazon-braket-sdk==1.18.0
numpy==1.22.2
boto3==1.21.30
